detec/untitled0.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
Created on Thu Apr 28 15:01:03 2022

@author: sajid
"""
from top2vec import Top2Vec
from sklearn.datasets import fetch_20newsgroups

newsgroups = fetch_20newsgroups(subset='all', remove=('headers', 'footers', 'quotes'))

model = Top2Vec(documents=newsgroups.data, speed="learn", workers=8)

topic_words, word_scores, topic_scores, topic_nums = model.search_topics(keywords=["finance"], num_topics=5)