detec/requirementsbis.txt
click==8.1.3
cycler==0.11.0
flatbuffers==1.12
fonttools==4.33.3
gast==0.3.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
h5py==2.10.0
idna==3.3
importlib-metadata==4.11.4
joblib==1.1.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
llvmlite==0.38.1
Markdown==3.3.7
matplotlib==3.5.2
nltk==3.6.5
numba==0.55.1
numpy==1.19.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
patsy==0.5.2
Pillow==9.1.1
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyod==0.9.9
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.15.0
SQLAlchemy==1.4.36
statsmodels==0.13.2
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==3.1.0
tqdm==4.64.0
typing-extensions==3.7.4.3
Unidecode==1.3.4
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.12.1
zipp==3.8.0
